refactor(st-09073): modularize update executor
